backend/requirements.txt
# Core dependencies
fastapi>=0.95.0,<0.100.0
uvicorn>=0.15.0,<0.22.0
python-multipart>=0.0.5
pytesseract>=0.3.8
Pillow>=8.0.0,<10.0.0
pymupdf>=1.19.0
sqlalchemy>=1.4.0,<2.0.0
psycopg2-binary>=2.8.6  # PostgreSQL driver
alembic>=1.8.0  # Database migrations
numpy>=1.19.0,<1.22.0

# AI/ML dependencies
torch>=1.13.0,<2.0.0
transformers>=4.18.0,<4.30.0
spacy>=3.2.0,<3.5.0
scikit-learn>=0.24.0,<1.0.0 